requirements.txt
wxPython==4.2.1
Pypubsub==4.0.3
